pyproject.toml
# pyproject.toml for Python dependencies within the Go agent_runtime project

[build-system]
requires = ["uv"]
build-backend = "uv.build"

[project]
name = "agent_runtime_python_deps" # Placeholder name
version = "0.1.0" # Placeholder version
description = "Python dependencies for SWE-Agent components integrated into agent_runtime."
requires-python = ">=3.11"
license = { text = "MIT" } # Assuming MIT based on SWE-Agent, adjust if needed

# Core dependencies from SWE-Agent, Django backend, and ML infrastructure
dependencies = [
    # SWE-Agent dependencies
    "datasets",
    "numpy",
    "pandas",
    "rich",
    "ruamel.yaml",
    "tenacity",
    "unidiff",
    "simple-parsing",
    "rich-argparse",
    "flask",
    "flask-cors",
    "flask-socketio",
    "pydantic>=2.0.0", # Ensure v2+ for SWE-ReX compatibility
    "python-dotenv",
    "pydantic_settings",
    "litellm",
    "GitPython",
    "ghapi",
    "tabulate",
    "textual>=1.0.0",
    
    # SWE-ReX dependencies
    "fastapi",
    "uvicorn",
    "requests",
    "pexpect",
    "bashlex",
    "python-multipart",
    "boto3",
    "modal",
    
    # Django backend dependencies
    "Django==5.2",
    "djangorestframework==3.16.0",
    "django-cors-headers==4.7.0",
    "grpcio==1.71.0",
    "grpcio-tools==1.71.0",
    
    # Additional dependencies from requirements.txt
    "aiohttp>=3.8.0",
    "asyncio>=3.4.3",
    "python-dateutil>=2.8.2",
    "tqdm>=4.65.0",
]

[project.scripts]
sweagent = "sweagent.run.run:main"


[project.optional-dependencies]
# Development dependencies from SWE-Agent
dev = [
    "mike",
    "mkdocs-material",
    "mkdocs-glightbox",
    "mkdocs-include-markdown-plugin",
    "mkdocstrings[python]>=0.18",
    "pytest>=7.0.0",
    "pytest-cov",
    "pipx",
    "pre-commit",
    "pytest-xdist",
    "griffe-pydantic!=1.1.3",
    "pytest-asyncio>=0.21.0", # Added to handle async tests properly
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
# Add minimal setuptools config if needed, e.g., to find packages if installing
# packages = ["sweagent", "tools", "tests"] # Might be needed if installing via pip
include-package-data = true

[tool.setuptools.packages.find]
where = ["."] # Search for packages in the root directory
include = ["agent*", "tools*", "tests*", "agent_framework*"] # Include copied SWE-Agent and SWE-ReX dirs
namespaces = false

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function" # Explicitly set scope

markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "cloud: mark tests for modal/fargate/etc.",
]

[tool.uv]
exclude = ["tests"]
virtualenv = ".venv"
requirements = []
pth = ["backend/apps"]
python-path = ["$VIRTUAL_ENV/bin/python"]
editable = true
